Move shift_dismiss & shift_undismiss endpoints to ShiftController
