+++ src/AppBundle/Controller/ShiftController.php
@@ -9,8 +9,10 @@ use AppBundle\Event\ShiftBookedEvent;
 use AppBundle\Event\ShiftFreedEvent;
 use AppBundle\Event\ShiftValidatedEvent;
 use AppBundle\Event\ShiftInvalidatedEvent;
+use AppBundle\Event\ShiftDismissedEvent;
 use AppBundle\Form\ShiftType;
 use AppBundle\Security\MembershipVoter;
+use AppBundle\Security\ShiftVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,6 +20,7 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\Form\Extension\Core\Type\RadioType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
@@ -291,6 +294,82 @@ class ShiftController extends Controller
 
         $referer = $request->headers->get('referer');
         return new RedirectResponse($referer);
+    }
+
+    /**
+     * Dismiss a booked shift
+     *
+     * @Route("/{id}/dismiss", name="shift_dismiss")
+     * @Method("POST")
+     */
+    public function dismissShiftAction(Request $request, Shift $shift)
+    {
+        if (!$this->isGranted('dismiss', $shift)) {
+            $session = new Session();
+            $session->getFlashBag()->add("error", "Impossible d'annuler ce créneau");
+            return $this->redirectToRoute("booking");
+        }
+
+        $beneficiary = $shift->getShifter();
+        $em = $this->getDoctrine()->getManager();
+        if($shift->isFixe()) {
+            $session = new Session();
+            $session->getFlashBag()->add("error", "Impossible d'annuler un créneau fixe");
+            return $this->redirectToRoute("booking");
+        } else {
+            $shift->setShifter(null);
+            $shift->setBooker(null);
+            $shift->setFixe(false);
+        }
+        $em->persist($shift);
+        $em->flush();
+
+        $reason = $request->get("reason");
+        $dispatcher = $this->get('event_dispatcher');
+        $dispatcher->dispatch(ShiftDismissedEvent::NAME, new ShiftDismissedEvent($shift, $beneficiary, $reason));
+
+        $session->getFlashBag()->add('success', "La créneau a été annulé");
+        return $this->redirectToRoute('homepage');
+    }
+
+    /**
+     * Undismiss a shift
+     *
+     * @Route("/undismiss", name="shift_undismiss")
+     * @Method("POST")
+     */
+    public function undismissShift(Request $request)
+    {
+        $session = new Session();
+
+        $form = $this->createFormBuilder()
+            ->setAction($this->generateUrl('shift_undismiss'))
+            ->setMethod('POST')
+            ->add('shift_id', HiddenType::class)
+            ->getForm();
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $em = $this->getDoctrine()->getManager();
+            $shift_id = $form->get('shift_id')->getData();
+            $shift = $em->getRepository('AppBundle:Shift')->find($shift_id);
+            if ($shift) {
+                $shift->setIsDismissed(false);
+                $shift->setDismissedTime(null);
+                $shift->setDismissedReason(null);
+
+                $em->persist($shift);
+                $em->flush();
+
+                $dispatcher = $this->get('event_dispatcher');
+                $dispatcher->dispatch(ShiftBookedEvent::NAME, new ShiftBookedEvent($shift, false));
+            } else {
+                $session->getFlashBag()->add('warning', "shift not found");
+            }
+        }
+
+        return $this->redirectToRoute('homepage');
     }
 
     /**
